Reorganize handlers to separate mock directory for reuse.
